[SCRUM-110] rename: fixture로 이름 변경
