+++ src/test/java/com/projectlyrics/server/domain/artist/repository/ArtistQueryRepositoryTest.java
@@ -3,7 +3,7 @@ package com.projectlyrics.server.domain.artist.repository;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import com.projectlyrics.server.domain.artist.entity.Artist;
-import com.projectlyrics.server.utils.ArtistTestUtil;
+import com.projectlyrics.server.fixture.ArtistFixture;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
@@ -24,8 +24,8 @@ public class ArtistQueryRepositoryTest {
   @Test
   void 아티스트의_데이터를_사용자_입력_쿼리_기반으로_조회해_반환한다() {
     // given
-    Artist artist1 = ArtistTestUtil.createWithName("검정치마");
-    Artist artist2 = ArtistTestUtil.createWithName("구남과여라이딩스텔라");
+    Artist artist1 = ArtistFixture.createWithName("검정치마");
+    Artist artist2 = ArtistFixture.createWithName("구남과여라이딩스텔라");
     artistCommandRepository.save(artist1);
     artistCommandRepository.save(artist2);
