+++ src/test/java/com/projectlyrics/server/domain/artist/service/ArtistCommandServiceTest.java
@@ -15,7 +15,7 @@ import com.projectlyrics.server.domain.artist.repository.ArtistCommandRepository
 import com.projectlyrics.server.domain.artist.repository.ArtistQueryRepository;
 import com.projectlyrics.server.domain.common.entity.enumerate.EntityStatusEnum;
 import com.projectlyrics.server.global.exception.FeelinException;
-import com.projectlyrics.server.utils.ArtistTestUtil;
+import com.projectlyrics.server.fixture.ArtistFixture;
 import java.util.Optional;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
@@ -43,7 +43,7 @@ class ArtistCommandServiceTest {
   @Test
   void 전달받은_데이터로_새로운_아티스트를_추가한다() {
     // given
-    Artist artist = ArtistTestUtil.create();
+    Artist artist = ArtistFixture.create();
     ArtistAddRequest addArtistRequest = createAddArtistRequest();
     given(artistCommandRepository.save(any(Artist.class))).willReturn(artist);
 
@@ -62,7 +62,7 @@ class ArtistCommandServiceTest {
   void NULL_또는_공백_빈문자열이_아닌_데이터로만_아티스트의_데이터를_수정한다() {
     // given
     Long artistId = 1L;
-    Artist artist = ArtistTestUtil.create();
+    Artist artist = ArtistFixture.create();
     ArtistUpdateRequest updateArtistRequest = createUpdateArtistRequest("   ", null, "https://~2");
     given(artistQueryRepository.findByIdAndNotDeleted(artistId)).willReturn(Optional.of(artist));
 
@@ -80,7 +80,7 @@ class ArtistCommandServiceTest {
   void 아티스트_데이터_수정_시_profileImageCdnLink가_https로_시작하지_않는다면_에러가_발생한다() {
     // given
     Long artistId = 1L;
-    Artist artist = ArtistTestUtil.create();
+    Artist artist = ArtistFixture.create();
     ArtistUpdateRequest updateArtistRequest = createUpdateArtistRequest(null, null, "http://~2");
     given(artistQueryRepository.findByIdAndNotDeleted(artistId)).willReturn(Optional.of(artist));
 
@@ -96,7 +96,7 @@ class ArtistCommandServiceTest {
   void 아티스트_데이터_삭제_시_soft_delete를_시킨다() {
     // given
     Long artistId = 1L;
-    Artist artist = ArtistTestUtil.create();
+    Artist artist = ArtistFixture.create();
     given(artistQueryRepository.findByIdAndNotDeleted(artistId)).willReturn(Optional.of(artist));
 
     // when
