+++ src/test/java/com/projectlyrics/server/domain/record/service/RecordCommandServiceTest.java
@@ -12,9 +12,9 @@ import com.projectlyrics.server.domain.record.dto.request.RecordAddRequest;
 import com.projectlyrics.server.domain.record.repository.RecordCommandRepository;
 import com.projectlyrics.server.domain.user.entity.User;
 import com.projectlyrics.server.domain.user.service.UserQueryService;
-import com.projectlyrics.server.utils.ArtistTestUtil;
-import com.projectlyrics.server.utils.RecordTestUtil;
-import com.projectlyrics.server.utils.UserTestUtil;
+import com.projectlyrics.server.fixture.ArtistFixture;
+import com.projectlyrics.server.fixture.RecordFixture;
+import com.projectlyrics.server.fixture.UserFixture;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.ArgumentCaptor;
@@ -45,14 +45,14 @@ class RecordCommandServiceTest {
   void 전달받은_데이터로_새로운_레코드를_추가한다() {
     // given
     long artistId = 1L;
-    Artist artist = ArtistTestUtil.createWithName("검정치마");
+    Artist artist = ArtistFixture.createWithName("검정치마");
     given(artistQueryService.getArtistById(artistId)).willReturn(artist);
 
     long userId = 1L;
-    User user = UserTestUtil.create();
+    User user = UserFixture.create();
     given(userQueryService.getUserById(userId)).willReturn(user);
 
-    given(recordCommandRepository.save(any(Record.class))).willReturn(RecordTestUtil.create(user, artist));
+    given(recordCommandRepository.save(any(Record.class))).willReturn(RecordFixture.create(user, artist));
 
     // when
     sut.addRecord(userId, new RecordAddRequest(artistId));
